feat: adds linear conv and lstm test for onnx registration
